+++ skywalking/log/sw_logging.py
@@ -81,10 +81,15 @@ def install():
             traceContext=TraceContext(
                 traceId=str(context.segment.related_traces[0]),
                 traceSegmentId=str(context.segment.segment_id),
-                spanId=context.active_span().sid if context.active_span() else -1
+                spanId=context.active_span.sid if context.active_span else -1
             ),
             tags=build_log_tags(),
         )
+
+        primary_endpoint_name = context.primary_endpoint.get_name()
+        if primary_endpoint_name:
+            log_data.endpoint = primary_endpoint_name
+
         _handle(self=self, record=record)
 
         agent.archive_log(log_data)