+++ skywalking/plugins/sw_tornado.py
@@ -44,7 +44,7 @@ def install():
 
     def _sw_handler_uncaught_exception(self: RequestHandler, ty, value, tb, *args, **kwargs):
         if value is not None:
-            entry_span = get_context().active_span()
+            entry_span = get_context().active_span
             if entry_span is not None:
                 entry_span.raised()