+++ skywalking/utils/exception.py
@@ -18,3 +18,10 @@
 class VersionRuleException(Exception):
     def __init__(self, message):
         self.message = message
+
+
+class IllegalStateError(RuntimeError):
+    """
+    Raised when tracing context falls into unexpected state.
+    """
+    pass